Make use of refactored htcondor workflow
