+++ generation/framework/htcondor/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ['HTCondorWorkflow']
+
+from .HTCondor import HTCondorWorkflow
