+++ generation/tasks/SherpaRun.py
@@ -6,7 +6,8 @@ import random
 
 from generation.framework.utils import run_command
 
-from generation.framework.tasks import GenRivetTask, HTCondorWorkflow, GenerationScenarioConfig
+from generation.framework.tasks import GenRivetTask, GenerationScenarioConfig
+from generation.framework.htcondor import HTCondorWorkflow
 
 from .SherpaIntegrate import SherpaIntegrate
 from .SherpaBuild import SherpaConfig
